feat: add Advanced tab
